Swap SQLite to Postgres

+++ Tests/StravaFakeServer/Generators/EventDataRequestFaker.cs
@@ -7,8 +7,8 @@ public class EventDataRequestFaker : Faker<EventDataRequest>
 {
     public EventDataRequestFaker()
     {
-        RuleFor(x => x.ObjectId, x => (ulong)x.Random.Long());
-        RuleFor(x => x.EventTime, (ulong)DateTimeOffset.UtcNow.ToUnixTimeSeconds());
-        RuleFor(x => x.SubscriptionId, x => (ulong)x.Random.Long());
+        RuleFor(x => x.ObjectId, x => x.Random.Long());
+        RuleFor(x => x.EventTime, DateTimeOffset.UtcNow.ToUnixTimeSeconds());
+        RuleFor(x => x.SubscriptionId, x => x.Random.Long());
     }
 }
